+++ QuanLyTangThuHoKhau.Core/AppServices/HoSoCuTruServices/Types/HSCT.cs
@@ -4,13 +4,13 @@ namespace QuanLyTangThuHoKhau.Core.AppServices.HoSoCuTruServices.Types
 {
     public class HSCT
     {
-        public uint SoHSCT { get; set; }
+        public int SoHSCT { get; set; }
         public string ChuHo { get; set; }
         public string MaHSCTDayDu { get; private set; }
 
         public HSCT(uint soHSCT, ThonXom thonXom, string chuHo = "")
         {
-            SoHSCT = soHSCT;
+            SoHSCT = (int)soHSCT;
             ChuHo = chuHo;
             MaHSCTDayDu = $"{thonXom.DonViHanhChinhPhuongXa.MaDonVi}-{soHSCT:D6}";
         }
